Add files via upload
In this project, we will be detecting humans from any video fed as an input and it will be displaying the count of the human like objects detected by the algorithm present in the given video. Python language has been used to develop this project and the video used in the project has been extracted from the internet.